refactor: move config file inside comm
