+++ src/index.ts
@@ -1,7 +1,7 @@
 import express from 'express'
 import http from 'http'
 import { Server } from 'socket.io'
-import { standardConfig } from './infrastructure/config'
+import { standardConfig } from './infrastructure/communication/config'
 import { acceptCommands } from './infrastructure/acceptCommands'
 
 const app = express()
