+++ src/infrastructure/communication/externalServicesUtils.ts
@@ -1,6 +1,6 @@
 import axios, { AxiosResponse } from 'axios'
 import { IAuthServiceUtils, IProfileServiceUtils } from '../../domain/utils/serviceUtils'
-import { standardConfig } from '../config'
+import { standardConfig } from './config'
 
 function httpGet(
   hostname: string,
